+++ CHANGELOG.md
@@ -6,7 +6,7 @@
     - multixy: doubleTap not working on touch devices
     - button: `x` and `y` variables not passed to `onValue` when `doubleTap` is true on touch devices  
     - advanced syntax: update issue with clones using their own value in their properties
-
+    - client: minor keyboard shortcut issues
 
 ## 1.24.2